Changed the implementation of map to be a class that has the logic and bound it to the mapMutation
