--- loremaster/backend/src/repositories/mapRepo.js
@@ -1,10 +1,0 @@
-const { Insert, Select } = require('../sql/sqlFunctions');
-
-module.exports = {
-    CreateAccount: () => {
-        
-    },
-    GetAccountWhere: () => {
-
-    }
-}
