+++ loremaster/backend/src/sql/sqlFunctions.js
@@ -1,33 +1,32 @@
 const mysql = require('mysql');
 
-module.exports = {
-    Connect: () => {
-        // Get all of the info to connect to the db 
-        const connection = mysql.createConnection({
-            host: process.env.HOST,
-            database: process.env.DATABASE,
-            user: process.env.USER,
-            password: process.env.SQLPASSWORD
-        })
-        connection.connect(function(error){
-            // If we cannot connect to the db then we should have a way to handle that
-            if(error){
-                console.log(error);
-                throw new Error(error);                
-                //connection = null
-            }
-        })
-        return connection;
-    },
+function Connect() {
+    // Get all of the info to connect to the db 
+    const connection = mysql.createConnection({
+        host: process.env.HOST,
+        database: process.env.DATABASE,
+        user: process.env.USER,
+        password: process.env.SQLPASSWORD
+    })
+    connection.connect(function(error){
+        // If we cannot connect to the db then we should have a way to handle that
+        if(error){
+            console.log(error);
+            throw new Error(error);                
+        }
+    })
+    return connection;
+}
 
+module.exports = {
     Select: (table, expression = null) => {
         // Establish a connection with the MySQL Server
-        const connection = this.Connect();
+        const connection = Connect();
 
         // Eventually this will be logged for every query made to the db
         const query = `
-        SELECT * FROM ${table}
-        ${(expression) ? 'WHERE '+ expression : ``}`;
+            SELECT * FROM ${table}
+            ${(expression) ? 'WHERE '+ expression : ``}`;
 
         //Output what the user is doing, ----MAKE THIS ONLY OUTPUT WITH --verbose----
         console.log(query);
@@ -51,20 +50,25 @@ module.exports = {
         })
     },
     Insert: (table, keys, values) => {
-        const connection = this.Connect();
-        return new Promise((success, failure) => {
-            connection.query(`
+        const connection = Connect();
+
+        const query = `
             INSERT INTO ${table} 
-            (${keys.join()}) VALUES ("${values.join(`","`)}")`, 
+            (${keys.join()}) VALUES ("${values.join(`","`)}")`
+
+        console.log(query);
+
+        return new Promise((success, failure) => {
+            connection.query(query, 
             function(error) {
                 if (error) {
                     console.log(error);
-                    return failure({
+                    failure({
                         code: "400", 
                         message: error.message
                     });
                 } else {
-                    return success({
+                    success({
                         code: 200, 
                         message: "Successfully inserted data!"
                     });
